refactor: rename "api" module to "Store"
